fix : 데미지 텍스트 추가하던 중 오류 있어서 되돌림

+++ Assets/Scripts/Monster/Monster.cs
@@ -10,9 +10,10 @@ public abstract class Monster : MonoBehaviour,IDamageable
     private float initHp;
     public event UnityAction<float> onChangeHp;
 
+
    
 
-    public void HitDamage(float damage)
+    public void HitDamage(int damage)
     {
         hp -= damage;
         onChangeHp?.Invoke(hp);
