+++ Assets/Scripts/Player/Player.cs
@@ -8,7 +8,7 @@ public abstract class Player : MonoBehaviour,IDamageable
     protected float hp;
     protected float initHp;
     public event UnityAction<float> onChangeHp;
-    public void HitDamage(float damage)
+    public void HitDamage(int damage)
     {
         hp -= damage;
         onChangeHp?.Invoke(hp);
